% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSfunctions.R
\name{LS.model}
\alias{LS.model}
\title{Least Squares Regression}
\usage{
LS.model(formula, data = NULL)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{data frame to which the formula relates}
}
\value{
list containing three elements: Parameters, df, y
}
\description{
Least Squares Regression
}
\examples{
df = data.frame(y = c(1,2,3,4), x = c(2,5,3,1))
LS.model(y~x, data=df)
}
