% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSfunctions.R
\name{LS.p}
\alias{LS.p}
\title{Least Squares Prediction}
\usage{
LS.p(model, newdata = NULL)
}
\arguments{
\item{model}{output from \code{LS.r} function}

\item{newdata}{input data to predict new output}
}
\value{
In-sample prediction vector if \code{newdata} not supplied. Otherwise out-of-sample prediction given by input \code{newdata}
}
\description{
Least Squares Prediction
}
