% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSfunctions.R
\name{LS.predict}
\alias{LS.predict}
\title{Least Squares Prediction}
\usage{
LS.predict(model, newdata = NULL)
}
\arguments{
\item{model}{output from \code{LS.r} function}

\item{newdata}{input data to predict new output}
}
\value{
In-sample prediction vector if \code{newdata} not supplied. Otherwise out-of-sample prediction given by input \code{newdata}
}
\description{
Least Squares Prediction
}
\examples{
df = data.frame(y = c(1,2,3,4), x = c(2,5,3,1))
m = LS.model(y~x, data=df)
LS.predict(m)
}
