% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSfunctions.R
\name{LS.r}
\alias{LS.r}
\title{Least Squares Regression}
\usage{
LS.r(formula, data = NULL)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{data frame to which the formula relates}
}
\value{
list containing two elements: Parameters, df
}
\description{
Least Squares Regression
}
\examples{
library(lasso2)
data(Prostate)
LS.r(lpsa ~ lcavol, data = Prostate)
}
